Add files via upload
The "Login Simulator" challenge on Hack The Box is similar to an online game that simulates actual scenarios in which we assume the position of a cyber investigator.  Usually, when we faced with this difficulty, our job is to navigate a login page or system in order to access regions that are prohibited. The difficulty frequently lies in locating holes in the authentication system, including weak passwords, improper coding techniques, or incorrect setups. In this project we use a variety of strategies, such as brute-force attacks, password guessing, and taking advantage of weaknesses in the login system.